Improved consistency between inference and training. Bug fix in Subdue. Parameter changes (debug params) for Caltech. Position calculation depends on previous level, not leaf nodes.

+++ 2D/CHOP/Prototype/graphTools/createEdgesWithLabels.m
@@ -137,7 +137,7 @@ function [mainGraph] = createEdgesWithLabels(mainGraph, options, currentLevelId)
         
         % Update data structures based on removed edges.
         allEdges = allEdges(validEdges,:);
-        edgeCoords = edgeCoords(validEdges,:);
+        edgeCoords = double(edgeCoords(validEdges,:));
         normalizedEdgeCoords = fix(fix(downsampleRatio * edgeCoords) + halfMatrixSize);
         
         % Double check in order not to go out of bounds.
